+++ Teams/ReferenceSolutionAStar/.coveragerc
@@ -1,0 +1,8 @@
+[run]
+omit =
+    # omit anything in a .local directory anywhere
+    */test_*
+    */__init__.py
+include = 
+    ./*
+
