+++ Teams/ReferenceSolutionAStar/result/UnitTestResult.txt
@@ -1,0 +1,3 @@
+Name                     Stmts   Miss  Cover   Missing
+------------------------------------------------------
+MazeSolverAlgoAStar.py     162     19    88%   26, 29, 32, 35, 38, 49-50, 65-67, 70, 85, 107, 111, 243-248
